feat: refactor ai action by splitting it into several files
