+++ ai/actions.tsx
@@ -1,37 +1,26 @@
+"use server";
+
 import { generateId, type CoreMessage, type ToolResultPart } from "ai";
 import {
-  createAI,
   createStreamableUI,
   createStreamableValue,
-  getAIState,
   getMutableAIState,
-  type StreamableValue,
 } from "ai/rsc";
-import { AnswerSection } from "@/components/answer-section";
-import { CopilotDisplay } from "@/components/copilot-display";
 import { ErrorCard } from "@/components/error-card";
 import { FollowupPanel } from "@/components/followup-panel";
-import { RetrieveSection } from "@/components/retrieve-section";
-import { SearchRelated } from "@/components/search-related";
-import { SearchSection } from "@/components/search-section";
 import { Section } from "@/components/section";
 import { Spinner } from "@/components/ui/spinner";
-import { UserMessage } from "@/components/user-message";
-import { VideoSearchSection } from "@/components/video-search-section";
-import { saveChat } from "@/lib/actions/chat";
 import { inquire, querySuggestor, researcher, taskManager } from "@/lib/agents";
 import { writer } from "@/lib/agents/writer";
-import type { AIMessage, Chat } from "@/lib/types";
 import { transformToolMessages } from "@/lib/utils/get-model";
+import type { AIMessage } from "./types";
 
-async function submit(
+export async function submit(
   formData?: FormData,
   skip?: boolean,
   retryMessages?: AIMessage[],
 ) {
-  "use server";
-
-  const aiState = getMutableAIState<typeof AI>();
+  const aiState = getMutableAIState();
   const uiStream = createStreamableUI();
   const isGenerating = createStreamableValue(true);
   const isCollapsed = createStreamableValue(false);
@@ -48,7 +37,7 @@ async function submit(
     )
     .map((message) => {
       const { role, content } = message;
-      return { role, content } as CoreMessage;
+      return { role, content };
     });
 
   // goupeiId is used to group the messages for collapse
@@ -90,16 +79,15 @@ async function submit(
         },
       ],
     });
-    messages.push({
-      role: "user",
-      content,
-    });
+    messages.push({ role: "user", content });
   }
 
   async function processEvents() {
     let action = { object: { next: "proceed" } };
     // If the user skips the task, we proceed to the search
-    if (!skip) action = (await taskManager(messages)) ?? action;
+    if (!skip) {
+      action = (await taskManager(messages)) ?? action;
+    }
 
     if (action.object.next === "inquire") {
       // Generate inquiry
@@ -255,194 +243,3 @@ async function submit(
     isCollapsed: isCollapsed.value,
   };
 }
-
-export type AIState = {
-  messages: AIMessage[];
-  chatId: string;
-  isSharePage?: boolean;
-};
-
-export type UIState = {
-  id: string;
-  component: React.ReactNode;
-  isGenerating?: StreamableValue<boolean>;
-  isCollapsed?: StreamableValue<boolean>;
-}[];
-
-const initialAIState: AIState = {
-  chatId: generateId(),
-  messages: [],
-};
-
-const initialUIState: UIState = [];
-
-// AI is a provider you wrap your application with so you can access AI and UI state in your components.
-export const AI = createAI<AIState, UIState>({
-  actions: {
-    submit,
-  },
-  initialUIState,
-  initialAIState,
-  onGetUIState: async () => {
-    "use server";
-
-    const aiState = getAIState();
-    if (aiState) {
-      const uiState = getUIStateFromAIState(aiState as Chat);
-      return uiState;
-    } else {
-      return;
-    }
-  },
-  onSetAIState: async ({ state, done }) => {
-    "use server";
-
-    // Check if there is any message of type 'answer' in the state messages
-    if (!state.messages.some((e) => e.type === "answer")) {
-      return;
-    }
-
-    const { chatId, messages } = state;
-    const createdAt = new Date();
-    const userId = "anonymous";
-    const path = `/search/${chatId}`;
-    const title =
-      messages.length > 0
-        ? JSON.parse(messages[0].content)?.input?.substring(0, 100) ||
-          "Untitled"
-        : "Untitled";
-    // Add an 'end' message at the end to determine if the history needs to be reloaded
-    const updatedMessages: AIMessage[] = [
-      ...messages,
-      {
-        id: generateId(),
-        role: "assistant",
-        content: `end`,
-        type: "end",
-      },
-    ];
-
-    const chat: Chat = {
-      id: chatId,
-      createdAt,
-      userId,
-      path,
-      title,
-      messages: updatedMessages,
-    };
-    await saveChat(chat);
-  },
-});
-
-export const getUIStateFromAIState = (aiState: Chat) => {
-  const chatId = aiState.chatId;
-  const isSharePage = aiState.isSharePage;
-  return aiState.messages
-    .map((message, index) => {
-      const { role, content, id, type, name } = message;
-
-      if (
-        !type ||
-        type === "end" ||
-        (isSharePage && type === "related") ||
-        (isSharePage && type === "followup")
-      )
-        return null;
-
-      switch (role) {
-        case "user":
-          switch (type) {
-            case "input":
-            case "input_related":
-              const json = JSON.parse(content);
-              const value = type === "input" ? json.input : json.related_query;
-              return {
-                id,
-                component: (
-                  <UserMessage
-                    message={value}
-                    chatId={chatId}
-                    showShare={index === 0 && !isSharePage}
-                  />
-                ),
-              };
-            case "inquiry":
-              return {
-                id,
-                component: <CopilotDisplay content={content} />,
-              };
-          }
-        case "assistant":
-          const answer = createStreamableValue();
-          answer.done(content);
-          switch (type) {
-            case "answer":
-              return {
-                id,
-                component: <AnswerSection result={answer.value} />,
-              };
-            case "related":
-              const relatedQueries = createStreamableValue();
-              relatedQueries.done(JSON.parse(content));
-              return {
-                id,
-                component: (
-                  <Section title="Related" separator={true}>
-                    <SearchRelated relatedQueries={relatedQueries.value} />
-                  </Section>
-                ),
-              };
-            case "followup":
-              return {
-                id,
-                component: (
-                  <Section title="Follow-up" className="pb-8">
-                    <FollowupPanel />
-                  </Section>
-                ),
-              };
-          }
-        case "tool":
-          try {
-            const toolOutput = JSON.parse(content);
-            const isCollapsed = createStreamableValue();
-            isCollapsed.done(true);
-            const searchResults = createStreamableValue();
-            searchResults.done(JSON.stringify(toolOutput));
-            switch (name) {
-              case "search":
-                return {
-                  id,
-                  component: <SearchSection result={searchResults.value} />,
-                  isCollapsed: isCollapsed.value,
-                };
-              case "retrieve":
-                return {
-                  id,
-                  component: <RetrieveSection data={toolOutput} />,
-                  isCollapsed: isCollapsed.value,
-                };
-              case "videoSearch":
-                return {
-                  id,
-                  component: (
-                    <VideoSearchSection result={searchResults.value} />
-                  ),
-                  isCollapsed: isCollapsed.value,
-                };
-            }
-          } catch (error) {
-            return {
-              id,
-              component: null,
-            };
-          }
-        default:
-          return {
-            id,
-            component: null,
-          };
-      }
-    })
-    .filter((message) => message !== null) as UIState;
-};
