+++ components/chat-panel.tsx
@@ -6,7 +6,8 @@ import { generateId } from "ai";
 import { useActions, useAIState, useUIState } from "ai/rsc";
 import { ArrowRight, Plus } from "lucide-react";
 import Textarea from "react-textarea-autosize";
-import type { AI, UIState } from "@/app/actions";
+import { AI } from "@/ai/provider";
+import type { UIState } from "@/ai/types";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { useAppState } from "@/lib/utils/app-state";
